+++ mcpp/mcpp3/ex01/ScavTrap.cpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ScavTrap.cpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: memartin <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/04 18:03:04 by memartin          #+#    #+#             */
-/*   Updated: 2020/10/21 17:29:24 by memartin         ###   ########.fr       */
+/*   Updated: 2020/11/24 15:20:24 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -128,7 +128,7 @@ ScavTrap::beRepaired(unsigned int amount)
 {
 	int heal;
 
-	if (amount > (this->_hpMax - this->_hp))
+	if (amount > (unsigned int)(this->_hpMax - this->_hp))
 		heal = this->_hpMax - this->_hp;
 	else
 		heal = amount;
